+++ src/caselawclient/Client.py
@@ -984,3 +984,35 @@ class MarklogicApiClient:
         )
 
         return results
+
+    def get_highest_parser_version(self) -> tuple[int, int]:
+        """This gets the highest parser version in the database, so if nothing has been parsed with the most recent version of the parser, this won't reflect that change."""
+        table = json.loads(
+            get_single_string_from_marklogic_response(
+                self._send_to_eval(
+                    {},
+                    "get_highest_parser_version.xqy",
+                )
+            )
+        )
+
+        return (int(table[1][1]), int(table[1][2]))
+
+    def get_pending_parse_for_version(
+        self, target_version: tuple[int, int]
+    ) -> list[list[Any]]:
+        """Retrieve documents which are not yet parsed with a given version."""
+        vars: query_dicts.GetPendingParseForVersionDict = {
+            "target_major_version": target_version[0],
+            "target_minor_version": target_version[1],
+        }
+        results: list[list[Any]] = json.loads(
+            get_single_string_from_marklogic_response(
+                self._send_to_eval(
+                    vars,
+                    "get_pending_parse_for_version.xqy",
+                )
+            )
+        )
+
+        return results