+++ src/caselawclient/xquery_type_dicts.py
@@ -83,6 +83,12 @@ class GetPendingEnrichmentForVersionDict(MarkLogicAPIDict):
     target_version: int
 
 
+# get_pending_parse_for_version.xqy
+class GetPendingParseForVersionDict(MarkLogicAPIDict):
+    target_major_version: int
+    target_minor_version: int
+
+
 # get_properties_for_search_results.xqy
 class GetPropertiesForSearchResultsDict(MarkLogicAPIDict):
     uris: list[Any]